Poll all zones for status on startup
This will help us notice if any zones are faulted at the time of
restart, like after a power outage and the bad guys have already
broken in. This re-uses the same slow inactivity loop as provided
for grabbing zone names, so it takes several seconds to fetch many
zones, but should help us heal our view of the world.

Related to issue #65